A trivial sample glTF parser and renderer that can render hello-world.gltf
This example simply shows the mechanics of parsing a glTF JSON file and
fetching buffer data from a data URI and then converting that parsed
data into OpenGL ES render commands. This is for the purposes of a
tutorial of how glTF works and is by far not a complete glTF renderer.

This example renders the hello-world.gltf scene (a triangle
using default material) into to a TextureView.